Refactored skb_roll submodule.
