+++ python/dicebeard/skb_roll/__init__.py
@@ -1,0 +1,3 @@
+from .beardeddie import *
+from .beardedroll import *
+from .skb_roll import *
